+++ functions/src/actions/_high-order-handlers/in-one-go.js
