+++ functions/src/actions/_high-order-handlers/middlewares/feeder-from-slots-scheme.js
