+++ functions/src/actions/_high-order-handlers/middlewares/next-song.js
