+++ functions/src/actions/_high-order-handlers/middlewares/play-song.js
