+++ functions/src/actions/_high-order-handlers/middlewares/prompt.js
