+++ functions/src/actions/_high-order-handlers/middlewares/render-speech.js
