+++ functions/src/actions/_high-order-handlers/middlewares/repair-broken-slots.js
