+++ functions/src/actions/in-one-go-music-playback.js
@@ -2,6 +2,6 @@ const playlist = require('../state/playlist');
 const query = require('../state/query');
 const strings = require('../strings').intents.inOneGoMusicPlayback;
 
-const inOneGo = require('./high-order-handlers/in-one-go');
+const inOneGo = require('./_high-order-handlers/in-one-go');
 
 module.exports = inOneGo.build({playlist, strings, query});
