+++ functions/tests/actions/_high-order-handlers/middlewares/ask.spec.js
@@ -1,7 +1,7 @@
 const {expect} = require('chai');
 const rewire = require('rewire');
 
-const middleware = rewire('../../../../src/actions/high-order-handlers/middlewares/ask');
+const middleware = rewire('../../../../src/actions/_high-order-handlers/middlewares/ask');
 
 const mockApp = require('../../../_utils/mocking/platforms/app');
 const mockDialog = require('../../../_utils/mocking/dialog');
