+++ functions/tests/actions/_high-order-handlers/middlewares/feeder-from-playlist.spec.js
@@ -1,7 +1,8 @@
 const {expect} = require('chai');
 const rewire = require('rewire');
 
-const middleware = rewire('../../../../src/actions/high-order-handlers/middlewares/feeder-from-slots-scheme');
+const feederFromPlaylist = rewire('../../../../src/actions/_high-order-handlers/middlewares/feeder-from-playlist');
+const playlist = require('../../../../src/state/playlist');
 
 const mockApp = require('../../../_utils/mocking/platforms/app');
 const mockFeeders = require('../../../_utils/mocking/feeders');
@@ -11,9 +12,6 @@ describe('actions', () => {
     let app;
     const feeder = {};
     let feeders;
-    const slotScheme = {
-      fulfilment: 'feederName',
-    };
 
     beforeEach(() => {
       app = mockApp();
@@ -21,16 +19,17 @@ describe('actions', () => {
       feeders = mockFeeders({
         getByNameReturn: feeder,
       });
-      middleware.__set__('feeders', feeders);
+      feederFromPlaylist.__set__('feeders', feeders);
     });
 
     describe('feeder from playlist', () => {
       it('should return Promise', () => {
-        expect(middleware()({app, slotScheme})).to.have.property('then');
+        expect(feederFromPlaylist.middleware()({app, playlist})).to.have.property('then');
       });
 
       it('should put feeder in context', () => {
-        return middleware()({app, slotScheme})
+        playlist.setFeeder(app, feeder);
+        return feederFromPlaylist.middleware()({app, playlist})
           .then(context => {
             expect(context).to.have.property('feeder', feeder);
           });
