+++ functions/tests/actions/_high-order-handlers/middlewares/feeder-from-slots-scheme.spec.js
@@ -1,8 +1,7 @@
 const {expect} = require('chai');
 const rewire = require('rewire');
 
-const feederFromPlaylist = rewire('../../../../src/actions/high-order-handlers/middlewares/feeder-from-playlist');
-const playlist = require('../../../../src/state/playlist');
+const middleware = rewire('../../../../src/actions/_high-order-handlers/middlewares/feeder-from-slots-scheme');
 
 const mockApp = require('../../../_utils/mocking/platforms/app');
 const mockFeeders = require('../../../_utils/mocking/feeders');
@@ -12,6 +11,9 @@ describe('actions', () => {
     let app;
     const feeder = {};
     let feeders;
+    const slotScheme = {
+      fulfilment: 'feederName',
+    };
 
     beforeEach(() => {
       app = mockApp();
@@ -19,17 +21,16 @@ describe('actions', () => {
       feeders = mockFeeders({
         getByNameReturn: feeder,
       });
-      feederFromPlaylist.__set__('feeders', feeders);
+      middleware.__set__('feeders', feeders);
     });
 
     describe('feeder from playlist', () => {
       it('should return Promise', () => {
-        expect(feederFromPlaylist.middleware()({app, playlist})).to.have.property('then');
+        expect(middleware()({app, slotScheme})).to.have.property('then');
       });
 
       it('should put feeder in context', () => {
-        playlist.setFeeder(app, feeder);
-        return feederFromPlaylist.middleware()({app, playlist})
+        return middleware()({app, slotScheme})
           .then(context => {
             expect(context).to.have.property('feeder', feeder);
           });
