+++ functions/tests/actions/_high-order-handlers/middlewares/fulfil-resolvers.spec.js
@@ -2,7 +2,7 @@ const {expect} = require('chai');
 const rewire = require('rewire');
 const sinon = require('sinon');
 
-const middleware = rewire('../../../../src/actions/high-order-handlers/middlewares/fulfil-resolvers');
+const middleware = rewire('../../../../src/actions/_high-order-handlers/middlewares/fulfil-resolvers');
 
 const mockTemplateResolvers = require('../../../_utils/mocking/template-resolver');
