+++ functions/tests/actions/_high-order-handlers/repair-handler-builder.spec.js
@@ -1,7 +1,7 @@
 const {expect} = require('chai');
 const rewire = require('rewire');
 
-const builder = rewire('../../../src/actions/high-order-handlers/repair-handler-builder');
+const builder = rewire('../../../src/actions/_high-order-handlers/repair-handler-builder');
 const {storeAction} = require('../../../src/state/actions');
 const strings = require('../../../src/strings').intents.noInput;
